Move unit tests to namespaces
