+++ Ctlg.UnitTests/Tests/Commands/FindCommandTests.cs
@@ -1,13 +1,12 @@
 ﻿using System;
 using Ctlg.Core;
 using Ctlg.Core.Interfaces;
-using Ctlg.Service;
 using Ctlg.Service.Commands;
 using Ctlg.Service.Utils;
 using Moq;
 using NUnit.Framework;
 
-namespace Ctlg.UnitTests
+namespace Ctlg.UnitTests.Tests.Commands
 {
     [TestFixture]
     public class FindCommandTests
