+++ Ctlg.UnitTests/Tests/Commands/Steps/SnapshotReaderTests.cs
@@ -4,12 +4,12 @@ using System.Linq;
 using Autofac.Extras.Moq;
 using Ctlg.Core;
 using Ctlg.Core.Interfaces;
-using Ctlg.Service;
 using Ctlg.Service.Commands;
+using Ctlg.UnitTests.Fixtures;
 using Moq;
 using NUnit.Framework;
 
-namespace Ctlg.UnitTests
+namespace Ctlg.UnitTests.Tests.Commands.Steps
 {
     public class SnapshotReaderTests : AutoMockTestFixture
     {
