+++ Ctlg.UnitTests/Tests/FileSizeTests.cs
@@ -2,7 +2,7 @@
 using Ctlg.Service.Utils;
 using NUnit.Framework;
 
-namespace Ctlg.UnitTests
+namespace Ctlg.UnitTests.Tests
 {
     [TestFixture]
     public class FileSizeTests
