+++ Ctlg.UnitTests/Tests/FileTests.cs
@@ -2,7 +2,7 @@
 using Ctlg.Core;
 using NUnit.Framework;
 
-namespace Ctlg.UnitTests
+namespace Ctlg.UnitTests.Tests
 {
     [TestFixture]
     public class FileTests
