+++ Ctlg.UnitTests/Tests/HashTests.cs
@@ -1,7 +1,7 @@
 ﻿using Ctlg.Core;
 using NUnit.Framework;
 
-namespace Ctlg.UnitTests
+namespace Ctlg.UnitTests.Tests
 {
     [TestFixture]
     public class HashTests
