+++ Ctlg.UnitTests/Tests/Services/IndexFileServiceTests.cs
@@ -3,12 +3,12 @@ using System.IO;
 using Autofac;
 using Autofac.Extras.Moq;
 using Ctlg.Core.Interfaces;
-using Ctlg.Service;
 using Ctlg.Service.Services;
+using Ctlg.UnitTests.Fixtures;
 using Moq;
 using NUnit.Framework;
 
-namespace Ctlg.UnitTests
+namespace Ctlg.UnitTests.Tests.Services
 {
     public class IndexFileServiceTests : AutoMockTestFixture
     {
